requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
boto3
python-dotenv
fastapi
dotenv
docutils==0.21.2
docx2txt==0.8
duckduckgo_search==6.3.3
email_validator==2.2.0
faiss-cpu==1.9.0
huggingface-hub==0.25.1
langchain==0.3.4
langchain-huggingface==0.1.0
langchain-ollama==0.2.0
langchain-openai==0.2.3
langchain-community
langgraph==0.2.43
nltk==3.9.1
PyMuPDF==1.24.12
PyMySQL==1.1.1
PyPDF2==3.0.1
tavily-python==0.5.0
textblob==0.18.0.post0
wikipedia==1.4.0
youtube-transcript-api==0.6.2
docling==2.15.1
ollama==0.4.6
openai==1.59.8

